Refactor tests to be based on language features instead of CST nodes
