+++ src/parser/__tests__/topLevelEdgeCases.ts
@@ -1,9 +1,19 @@
 import { parse } from '../parser';
 import { type Program } from '../../ast/types';
 
-describe('compilationUnit', () => {
+describe('Top-level edge cases', () => {
   it('returns an empty Program for the empty string', () => {
     const code = '';
+    const ast = parse(code);
+    const expectedAst: Program = {
+      type: 'Program',
+      body: []
+    };
+    expect(ast).toEqual(expectedAst);
+  });
+
+  it('filters out the empty statement', () => {
+    const code = ';';
     const ast = parse(code);
     const expectedAst: Program = {
       type: 'Program',
